model/etl/requirements.txt
elasticsearch
pandas
openpyxl
numpy
python-dateutil
python-dotenv